requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
colorama==0.4.6
distlib==0.3.7
Django==4.2.7
django-autoslug==1.9.9
django-bootstrap-v5==1.0.11
django-bootstrap4==23.2
django-bootstrap5==23.3
django-cities==0.6.1
django-cities-light==3.10.0
django-countries==7.5.1
django-filter==23.4
django-model-utils==4.3.1
django-phone-field==1.8.1
djangorestframework==3.14.0
eval==0.0.5
filelock==3.13.1
Markdown==3.5.1
Pillow==10.1.0
platformdirs==3.11.0
progressbar2==4.2.0
PyMonCtl==0.7
python-utils==3.8.1
pytube==15.0.0
pytz==2023.3.post1
pywin32==306
PyWinBox==0.6
PyWinCtl==0.3
setuptools==68.2.2
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
swapper==1.3.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
Unidecode==1.3.7
virtualenv==20.24.6
wheel==0.41.3
